chore: Update editor and terminal configurations
Update the editor and terminal configurations in the .config/Code/User/settings.json and .config/zsh/.zshrc files. Increase the font size in the editor to 16 and change the bat theme in the zshrc file to gruvbox-dark. Also, add the pyenv configuration to the zshrc file for managing Python versions.